[ivp_utility] x64 ready memory allocators

+++ havana/havok/hk_base/string/string.h
@@ -3,9 +3,9 @@
 
 class hk_String {
 public:
-	static int strcmp( const char *, const char * );
+	static int strcmp(const char *, const char * );
 	static void strcpy(char *, const char *);
-	static void memcpy(void *dest, const void *source, int size);
+	static void memcpy(void *dest, const void *source, size_t size);
 
 };
 typedef hk_String hkString;
